Switch ical/vcard/cert4android to our own repos.
DAVdroid have updated to kotlin, and kotlin is problematic with
annotation processors which was breaking the whole build.

This also meant we were not able to update to a new gradle version or
import just partial, smaller fixes. Now we have control over the
submodules as well, which fixes the above issues. At least for now.